autorisation en front uniqument pour le createur en update et delete

+++ backend/routes/posts.js
@@ -60,7 +60,7 @@ router.post(
     });
   }
 );
-/* Put mise a  jour d'un poste route protéger par le token via checkAuth*/
+/* Put mise a  jour d'un poste route protéger par le token via checkAuth seul le createur update*/
 router.put(
   "/:id",
   checkAuth,
@@ -78,8 +78,12 @@ router.put(
       imagePath: imagePath
     });
     console.log(post);
-    Post.updateOne({ _id: req.params.id }, post).then(result => {
-      res.status(200).json({ message: "Update successful!" });
+    Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+      if (result.nModified > 0) {
+        res.status(200).json({ message: "Update successful!" });
+      } else {
+      res.status(401).json({ message: "you are not the creator of the post not authorised!" });
+      }
     });
   }
 );
@@ -122,9 +126,14 @@ router.get("/:id", (req, res, next) => {
 
 /*REQUETE delete FOR POST via le parametre dynamique id route protéger par le token via checkAuth*/
 router.delete("/:id",checkAuth, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then(result => {
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
     console.log(result);
-    res.status(200).json({ message: "Post deleted!" });
+    if (result.n > 0) {
+      res.status(200).json({ message: "Post deleted!" });
+    } else {
+    res.status(401).json({ message: "you are not the creator of the post not authorised!" });
+    }
+
   });
 });
